requirements.txt
discord.py>=2.4.0
python-telegram-bot<13.0
tqdm
python-dotenv
